+++ lib/action/core/flow/callbacks.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require "action/core/flow/handlers"
+require "action/core/flow/handlers/resolvers/callback_resolver"
 
 module Action
   module Core
@@ -17,9 +18,13 @@ module Action
         module ClassMethods
           # Internal dispatcher
           def _dispatch_callbacks(event_type, action:, exception: nil)
-            _callbacks_registry.for(event_type).each do |handler|
-              handler.apply(action:, exception:)
-            end
+            resolver = Action::Core::Flow::Handlers::Resolvers::CallbackResolver.new(
+              _callbacks_registry,
+              event_type,
+              action:,
+              exception:,
+            )
+            resolver.execute_callbacks
           end
 
           # ONLY raised exceptions (i.e. NOT fail!).
@@ -39,13 +44,14 @@ module Action
 
           def _add_callback(event_type, handler: nil, block: nil, **kwargs)
             raise ArgumentError, "on_#{event_type} cannot be called with both :if and :unless" if kwargs.key?(:if) && kwargs.key?(:unless)
-
-            condition = kwargs.key?(:if) ? kwargs[:if] : kwargs[:unless]
+            raise ArgumentError, "on_#{event_type} cannot be called with both a block and a handler" if block && handler
             raise ArgumentError, "on_#{event_type} must be called with a block or symbol" unless block || handler
 
-            callback_handler = block || handler
-            matcher = condition.nil? ? nil : Action::Core::Flow::Handlers::Matcher.new(condition, invert: kwargs.key?(:unless))
-            entry = Action::Core::Flow::Handlers::CallbackHandler.new(matcher:, handler: callback_handler)
+            condition = kwargs.key?(:if) ? kwargs[:if] : kwargs[:unless]
+            matcher = condition ? Action::Core::Flow::Handlers::Matcher.new(condition, invert: kwargs.key?(:unless)) : nil
+            handler ||= block
+
+            entry = Action::Core::Flow::Handlers::Descriptors::CallbackDescriptor.new(matcher:, handler:)
             self._callbacks_registry = _callbacks_registry.register(event_type:, entry:)
           end
         end