+++ spec/action/core/messages_spec.rb
@@ -506,7 +506,10 @@ RSpec.describe Action do
               build_action do
                 success "Great news!", if: :condition?, unless: :other_condition?
               end
-            end.to raise_error(ArgumentError, /success cannot be called with both :if and :unless/)
+            end.to raise_error(Action::UnsupportedArgument,
+                               "calling success with both :if and :unless is not currently supported.\n\n" \
+                               "Implementation is technically possible but very complex. Please submit a " \
+                               "Github Issue if you have a real-world need for this functionality.")
           end
 
           it "raises ArgumentError for error" do
@@ -514,7 +517,10 @@ RSpec.describe Action do
               build_action do
                 error "Bad news!", if: :condition?, unless: :other_condition?
               end
-            end.to raise_error(ArgumentError, /error cannot be called with both :if and :unless/)
+            end.to raise_error(Action::UnsupportedArgument,
+                               "calling error with both :if and :unless is not currently supported.\n\n" \
+                               "Implementation is technically possible but very complex. Please submit a " \
+                               "Github Issue if you have a real-world need for this functionality.")
           end
         end
       end
@@ -534,16 +540,16 @@ RSpec.describe Action do
         it { expect(result).not_to be_ok }
         it { is_expected.to eq("Bad news!") }
 
-        it "supports class level default_error" do
-          expect(action.default_error).to eq("Bad news!")
+        it "supports result level default_error" do
+          expect(result.default_error).to eq("Bad news!")
         end
 
-        it "supports class level default_success" do
-          klass = build_action do
+        it "supports result level default_success" do
+          success_result = build_action do
             success "Great news!"
-          end
+          end.call
 
-          expect(klass.default_success).to eq("Great news!")
+          expect(success_result.default_success).to eq("Great news!")
         end
       end
 
@@ -565,8 +571,8 @@ RSpec.describe Action do
           is_expected.to eq("Bad news: ")
         end
 
-        it "supports class level default_error" do
-          expect(action.default_error).to eq("Bad news: ")
+        it "supports result level default_error" do
+          expect(result.default_error).to eq("Bad news: ")
         end
       end
 
@@ -603,8 +609,8 @@ RSpec.describe Action do
           is_expected.to eq("Bad news: Action::InboundValidationError")
         end
 
-        it "supports class level default_error" do
-          expect(action.default_error).to eq("Bad news: Action::Failure")
+        it "supports result level default_error" do
+          expect(result.default_error).to eq("Bad news: Action::InboundValidationError")
         end
 
         context "when fail! is called with custom message" do
@@ -707,8 +713,8 @@ RSpec.describe Action do
           is_expected.to eq("Something went wrong")
         end
 
-        it "supports class level default_error" do
-          expect(action.default_error).to eq("Something went wrong")
+        it "supports result level default_error" do
+          expect(result.default_error).to eq("Something went wrong")
         end
       end
 
@@ -767,7 +773,7 @@ RSpec.describe Action do
       end
 
       it "can reference configured error" do
-        expect(action.call(param: 4).error).to eq("whoa: Bad news!")
+        expect(action.call(param: 4).error).to eq("Bad news!")
       end
 
       it "falls back correctly" do